# BikeWorldServer/api-docs.yaml
openapi: 3.0.0
info:
  title: bike-world API
  version: '2.0'
servers:
  - url: https://bike-world-ids.herokuapp.com/
paths:
...
  /api/v2/itineraries/{itineraryId}/reviews:
    get:
      summary: Get Reviews
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: Itinerary Id 
          schema:
            type: string
            example: 627fb874a991e0a164e063ac
      responses:
        '200':
          description: List of the reviews.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id: 
                      type: string
                      example: "627fb874a991e0a164e063ac"
                    title:
                      type: string
                      example: 'Bellissimo'
                    text:
                      type: string
                      example: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.'
                    stars: 
                      type: integer
                      example: 4
                    author:
                      type: string
                      example: 'user'
                    self:
                      type: string
                      example: "/api/v2/itineraries/627fb874a991e0a164e063ac/reviews/627fb874a991e0a164e063ac"                
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'
        '404':
          description: Itinerary with id {itineraryId} not found
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Itinerary with id: {itineraryId} not found.'
...
  /api/v2/itineraries/{itineraryId}/reviews/{reviewId}:
    put:
      summary: Update a Review.
      parameters:
        - name: itineraryId
          in: path
          required: true
          description: Itinerary Id 
          schema:
            type: string
            example: 627fb874a991e0a164e063ac
        - name: reviewId
          in: path
          required: true
          description: Review Id 
          schema:
            type: string
            example: 627fb874a991e0a164e063ac
        - name: x-access-token
          in: header
          required: true
          description: The jwt of the session.
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJ0b3B0YWwuY29tIiwiZXhwIjoxNDI2NDIwODAwLCJodHRwOi8vdG9wdGFsLmNvbS9qd3RfY2xhaW1zL2lzX2FkbWluIjp0cnVlLCJjb21wYW55IjoiVG9wdGFsIiwiYXdlc29tZSI6dHJ1ZX0.yRQYnWzskCZUxPwaQupWkiUzKELZ49eM7oWxAQK_ZXw'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Bellissimo"
                text:
                  type: string
                  example: "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua."
                stars:
                  type: integer
                  example: 4
      responses:
        '200':
          description: Review updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: 
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Review info updated!"
        '404':
          description: Itinerary with id {itineraryId} not found
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Itinerary with id: {itineraryId} not found.'
        '404':
          description: Review with id {reviewId} not found
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Review with id: {reviewId} not found.'
        '401':
          description: Unauthorized. No token provided.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'No token provided.'
        '403':
          description: Forbidden. Failed to authenticate token.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Failed to authenticate token.'
        '401':
          description: Unauthorized. You can access only your reviews.
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Unauthorized. You can access only your reviews.'
        '400':
          description: Bad Request. Check docs for required parameters. /api/v1/api-docs
          content:
            application/json:
              schema: 
                type: object
                properties:
                  success: 
                    type: boolean
                    example: false
                  message: 
                    type: string
                    example: 'Bad Request. Check docs for required parameters. /api/v1/api-docs'